Allow registering globals from plugins
