+++ pyolin/plugins/__init__.py
@@ -1,0 +1,12 @@
+"""
+Each plugin should have a function with the following signature:
+
+    def register(
+        plugin_reg: PluginRegistration,
+        input_stream: Callable[[], ContextManager[typing.BinaryIO]],
+        config: PyolinConfig,
+    )
+"""
+from . import auto_parser, json
+
+PLUGINS = [auto_parser, json]
