+++ pyolin/pyolin.py
@@ -1,174 +1,48 @@
 """Main entry point for Pyolin, the utility to easily write Python one-liners."""
 
 import argparse
-from dataclasses import dataclass
 import importlib
-from io import TextIOWrapper
 import itertools
 import sys
-import json
 
 from contextlib import contextmanager
 from typing import (
     Any,
+    Callable,
+    ContextManager,
     Generator,
     Iterable,
-    Iterator,
     Optional,
     Tuple,
-    TypeVar,
     Union,
 )
 import typing
 from hashbang import command, Argument
 
+from pyolin.core import PluginRegistration, PyolinConfig
+
 from .field import DeferredType
 from .ioformat import (
     PARSERS,
     PRINTERS,
-    AbstractParser,
     Printer,
-    create_parser,
     gen_split,
     new_printer,
 )
-from .plugins import auto_parser, json as json_plugin
+from .plugins import PLUGINS
 from .util import (
     LazyItem,
     Item,
     ItemDict,
+    ReplayIter,
     StreamingSequence,
     _UNDEFINED_,
     NoMoreRecords,
-    peek_iter,
 )
-from .record import Header, RecordSequence
+from .record import RecordSequence
 from .parser import Prog
 
-
-json_plugin.register()
-auto_parser.register()
-
-
-@contextmanager
-def get_io(
-    input_: Union[typing.BinaryIO, str, None]
-) -> Generator[typing.BinaryIO, None, None]:
-    """Get the IO from the given input filename, or from stdin if `input_file`
-    is None."""
-    if isinstance(input_, str):
-        mode = "rb"
-        with open(input_, mode) as input_file:
-            yield input_file
-    elif input_ is not None:
-        yield input_
-    else:
-        yield sys.stdin.buffer
-
-
-T = TypeVar("T")
-_SENTINEL = object()
-
-
-class ReplayIter(Iterator[T]):
-
-    def __init__(self, iterator):
-        self._curval: Union[T, object] = _SENTINEL
-        self._iter = iterator
-
-    def __next__(self) -> T:
-        self._curval = next(self._iter)
-        return typing.cast(T, self._curval)
-
-    def current_or_first_value(self) -> T:
-        return typing.cast(T, self._curval) if self._curval is not _SENTINEL else next(self)
-
-    def has_multiple_items(self) -> bool:
-        preview, self._iter = peek_iter(self._iter, 2)
-        return len(preview) == 2
-
-
-@dataclass
-class ScopeIterator:
-    """An optional iterator that a pyolin program can set (typically via
-    accessing a provided variable), so that the program will continue executing
-    multiple times until the iterator is exhausted."""
-    iterator: Optional[Iterator[Any]]
-    # A name for the scope, for comparison and to display in error messages.
-    name: str
-
-
-class PyolinConfig:
-    """Configuration of Pyolin, available to the Pyolin program as `cfg`."""
-
-    _parser: Optional[AbstractParser] = None
-    _parser_frozen: bool = False
-    header: Optional[Header] = None
-    _scope_iterator: Optional[ScopeIterator] = None
-
-    def __init__(
-        self,
-        printer: Printer,
-        record_separator: str,
-        field_separator: Optional[str],
-        input_format: str,
-    ):
-        self.printer = printer
-        self._record_separator = record_separator
-        self._field_separator = field_separator
-        self._input_format = input_format
-
-    @property
-    def parser(self) -> AbstractParser:
-        """The parser for parsing input files, if `records` or `record` is used."""
-        if not self._parser:
-            self._parser = self.new_parser(self._input_format)
-        return self._parser
-
-    @parser.setter
-    def parser(self, value):
-        if self._parser_frozen:
-            raise RuntimeError("Parsing already started, cannot set parser")
-        if isinstance(value, str):
-            self._parser = self.new_parser(value)
-        elif isinstance(value, AbstractParser):
-            self._parser = value
-        else:
-            raise TypeError(
-                f"Expect `parser` to be an `AbstractParser`. Found `{value.__class__}` instead"
-            )
-
-    def new_parser(
-        self,
-        parser_format: str,
-        *,
-        record_separator: Optional[str] = None,
-        field_separator: Optional[str] = None,
-    ) -> AbstractParser:
-        """Create a new parser based on the given format and the current configuration."""
-        return create_parser(
-            parser_format,
-            record_separator or self._record_separator,
-            field_separator or self._field_separator,
-        )
-
-    def _freeze_parser(self) -> AbstractParser:
-        self._parser_frozen = True
-        return self.parser
-
-    def set_scope(self, scope_iterator: Optional[Iterator[Any]], name: str):
-        """Set the scope of the pyolin program execution.
-
-        The scope can only be set once per pyolin program. When set, pyolin will
-        execute the given program in a loop until the iterator is exhausted.
-        Therefore, the pyolin program and/or the registerer of this scope should
-        ensure that the iterator is advanced on every invocation."""
-        if self._scope_iterator is not None and self._scope_iterator.name != name:
-            raise RuntimeError(
-                f"Cannot change scope from "
-                f"\"{self._scope_iterator.name}\" to \"{name}\""
-            )
-        self._scope_iterator = ScopeIterator(scope_iterator, name)
+PLUGIN_REGISTRATION = PluginRegistration()
 
 
 def _execute_internal(
@@ -182,26 +56,42 @@ def _execute_internal(
 ) -> Tuple[Any, PyolinConfig]:
     prog = Prog(prog)
 
+    @contextmanager
+    def input_stream() -> Generator[typing.BinaryIO, None, None]:
+        """Get the IO from the given input filename, or from stdin if `input_file`
+        is None."""
+        if isinstance(input_, str):
+            mode = "rb"
+            with open(input_, mode) as input_file:
+                yield input_file
+        elif input_ is not None:
+            yield input_
+        else:
+            yield sys.stdin.buffer
+
     try:
         config = PyolinConfig(
-            new_printer(output_format), record_separator, field_separator, input_format
+            new_printer(output_format),
+            record_separator,
+            field_separator,
+            input_format,
         )
     except KeyError:
         raise ValueError(f'Unrecognized output format "{output_format}"') from None
 
-    def gen_records(input_file: Union[str, typing.BinaryIO, None]):
-        parser = config._freeze_parser()  # pylint:disable=protected-access
-        with get_io(input_file) as io_stream:
+    def gen_records(input_stream: Callable[[], ContextManager[typing.BinaryIO]]):
+        with input_stream() as io_stream:
+            parser = config._freeze_parser()  # pylint:disable=protected-access
             for i, record in enumerate(parser.records(io_stream)):
                 record.set_num(i)
                 yield record
 
-    def get_contents(input_file: Union[str, typing.BinaryIO, None]) -> DeferredType:
-        config._freeze_parser()  # pylint:disable=protected-access
-        with get_io(input_file) as io_stream:
+    def get_contents(input_stream: Callable[[], ContextManager[typing.BinaryIO]]) -> DeferredType:
+        with input_stream() as io_stream:
+            config._freeze_parser()  # pylint:disable=protected-access
             return DeferredType(io_stream.read())
 
-    record_seq = RecordSequence(gen_records(input_))
+    record_seq = RecordSequence(gen_records(input_stream))
 
     def get_dataframe():
         import pandas as pd  # pylint:disable=import-outside-toplevel
@@ -211,9 +101,7 @@ def _execute_internal(
         return dataframe.apply(pd.to_numeric, errors="ignore")  # type: ignore
 
     def file_scoped(func):
-        return LazyItem(
-            func, on_accessed=lambda: config.set_scope(None, "file")
-        )
+        return LazyItem(func, on_accessed=lambda: config.set_scope(None, "file"))
 
     iter_record_seq = ReplayIter(iter(record_seq))
 
@@ -224,11 +112,12 @@ def _execute_internal(
         except StopIteration:
             raise NoMoreRecords
 
-    def gen_lines():
-        with get_io(input_) as io_stream:
+    def gen_lines(input_stream: Callable[[], ContextManager[typing.BinaryIO]]):
+        with input_stream() as io_stream:
             for bytearr in gen_split(io_stream, "\n"):
-                yield bytearr.decode('utf-8')
-    iter_line_seq = ReplayIter(gen_lines())
+                yield bytearr.decode("utf-8")
+
+    iter_line_seq = ReplayIter(gen_lines(input_stream))
 
     def access_line_var():
         config.set_scope(iter_line_seq, "line")
@@ -237,41 +126,23 @@ def _execute_internal(
         except StopIteration:
             raise NoMoreRecords
 
-    def json_seq():
-        with get_io(input_) as io_stream:
-            text_stream = TextIOWrapper(io_stream)
-            finder = json_plugin.JsonFinder()
-            while line := text_stream.readline(1024):
-                yield from finder.add_input(line)
-
-    iter_json_seq = ReplayIter(json_seq())
-
-    def access_json():
-        if config._scope_iterator or iter_json_seq.has_multiple_items():
-            # Special case: Don't set scope if there is only one item, so that
-            # the output of a program using `jsonobj` will not present itself as
-            # a sequence.
-            config.set_scope(iter_json_seq, "json")
-        try:
-            return iter_json_seq.current_or_first_value()
-        except StopIteration:
-            raise NoMoreRecords
+    for plugin in PLUGINS:
+        plugin.register(PLUGIN_REGISTRATION, input_stream, config)
 
     global_dict = ItemDict(
         {
+            **PLUGIN_REGISTRATION._globals,
             # Record scoped
             "record": Item(access_record_var),
             "fields": Item(access_record_var),
             "line": Item(access_line_var),
-            "jsonobj": Item(access_json),
             # File scoped
             "lines": file_scoped(
                 lambda: StreamingSequence(r.source for r in record_seq)
             ),
-            "jsonobjs": file_scoped(json_seq),
             "records": file_scoped(lambda: record_seq),
-            "file": file_scoped(lambda: get_contents(input_)),
-            "contents": file_scoped(lambda: get_contents(input_)),
+            "file": file_scoped(lambda: get_contents(input_stream)),
+            "contents": file_scoped(lambda: get_contents(input_stream)),
             "df": file_scoped(get_dataframe),
             # Other
             "filename": input_,
@@ -292,9 +163,13 @@ def _execute_internal(
 
     try:
         result = prog.exec(global_dict)
-        if config._scope_iterator is not None and config._scope_iterator.iterator is not None:
+        if (
+            config._scope_iterator is not None
+            and config._scope_iterator.iterator is not None
+        ):
             result = itertools.chain(
-                (result,), (prog.exec(global_dict) for _ in config._scope_iterator.iterator)
+                (result,),
+                (prog.exec(global_dict) for _ in config._scope_iterator.iterator),
             )
         return result, config
     except NoMoreRecords:
