+++ pyolin/util.py
@@ -18,6 +18,7 @@ from typing import (
     Tuple,
     TypeVar,
     Union,
+    overload,
 )
 import typing
 
@@ -155,7 +156,7 @@ class Item(Generic[T]):
     """An Item that is defined by a function, which will be initialized when this item is used,
     typically from ItemDict."""
 
-    def __init__(self, func: Callable[..., T]):
+    def __init__(self, func: Callable[[], T]):
         self.func = func
 
     def __call__(self, *arg, **kwargs) -> T:
@@ -169,7 +170,7 @@ class LazyItem(Item[T]):
 
     def __init__(
         self,
-        func: Callable[..., T],
+        func: Callable[[], T],
         *,
         on_accessed: Optional[Callable[[], None]] = None,
     ):
@@ -185,6 +186,16 @@ class LazyItem(Item[T]):
                 self._on_accessed()
             self._cached = True
         return typing.cast(T, self._val)
+
+
+@overload
+def peek_iter(iterator: Sequence[T], num: int) -> Tuple[Sequence[T], Iterable[T]]:
+    ...
+
+
+@overload
+def peek_iter(iterator: Iterator[T], num: int) -> Tuple[Sequence[T], Iterator[T]]:
+    ...
 
 
 def peek_iter(iterator: Iterable[T], num: int) -> Tuple[Sequence[T], Iterable[T]]:
@@ -238,3 +249,25 @@ def clean_close_stdout_and_stderr() -> None:
                 sys.stderr.flush()
             finally:
                 sys.stderr.close()
+
+
+T = TypeVar("T")
+_SENTINEL = object()
+
+
+class ReplayIter(Iterator[T]):
+
+    def __init__(self, iterator: Iterator[T]):
+        self._curval: Union[T, object] = _SENTINEL
+        self._iter: Iterator[T] = iterator
+
+    def __next__(self) -> T:
+        self._curval = next(self._iter)
+        return typing.cast(T, self._curval)
+
+    def current_or_first_value(self) -> T:
+        return typing.cast(T, self._curval) if self._curval is not _SENTINEL else next(self)
+
+    def has_multiple_items(self) -> bool:
+        preview, self._iter = peek_iter(self._iter, 2)
+        return len(preview) == 2
